+++ src/accounts/safe.ts
@@ -1,6 +1,7 @@
 import {
   type Abi,
   type Address,
+  type Chain,
   concat,
   decodeFunctionData,
   encodeAbiParameters,
@@ -13,6 +14,7 @@ import {
   parseAbi,
   parseAbiParameters,
   zeroAddress,
+  zeroHash,
 } from 'viem'
 import {
   entryPoint07Abi,
@@ -36,6 +38,7 @@ import type { EnableSessionData } from '../modules/validators/smart-sessions'
 import type { OwnerSet, RhinestoneAccountConfig, Session } from '../types'
 import {
   AccountConfigurationNotSupportedError,
+  Eip712DomainNotAvailableError,
   OwnersFieldRequiredError,
 } from './error'
 import { encode7579Calls, getAccountNonce, type ValidatorConfig } from './utils'
@@ -161,6 +164,21 @@ function getAddress(config: RhinestoneAccountConfig) {
     bytecode: concat([SAFE_PROXY_INIT_CODE, constructorArgs]),
   })
   return address
+}
+
+function getEip712Domain(config: RhinestoneAccountConfig, chain: Chain) {
+  if (config.initData) {
+    throw new Eip712DomainNotAvailableError(
+      'Existing Safe-7579 accounts are not yet supported',
+    )
+  }
+  return {
+    name: 'rhinestone safe7579',
+    version: 'v1.0.0',
+    chainId: chain.id,
+    verifyingContract: getAddress(config),
+    salt: zeroHash,
+  }
 }
 
 function getInstallData(module: Module) {
@@ -369,6 +387,7 @@ function getOwners(config: RhinestoneAccountConfig) {
   const ownerSet = config.owners
   switch (ownerSet.type) {
     case 'ecdsa':
+    case 'ens':
       return ownerSet.accounts.map((account) => account.address)
     case 'passkey':
       return [NO_SAFE_OWNER_ADDRESS]
@@ -384,6 +403,7 @@ function getThreshold(config: RhinestoneAccountConfig) {
   const ownerSet = config.owners
   switch (ownerSet.type) {
     case 'ecdsa':
+    case 'ens':
       return ownerSet.threshold ? BigInt(ownerSet.threshold) : 1n
     case 'passkey':
       return 1n
@@ -393,6 +413,7 @@ function getThreshold(config: RhinestoneAccountConfig) {
 }
 
 export {
+  getEip712Domain,
   getInstallData,
   getAddress,
   packSignature,