+++ src/actions/smart-sessions.ts
@@ -1,10 +1,5 @@
-import { type Hex } from 'viem'
-import type { SessionDetails } from '../execution/smart-session'
-import {
-  encodeSmartSessionSignature as encodeSmartSessionSignatureInternal,
-  getEnableSessionCall,
-} from '../modules/validators/smart-sessions'
-import { Session } from '../types'
+import { getEnableSessionCall } from '../modules/validators/smart-sessions'
+import type { Session } from '../types'
 
 /**
  * Enable a smart session
@@ -19,22 +14,4 @@ function enableSession(session: Session) {
   }
 }
 
-/**
- * Encode a smart session signature
- * @param sessionDetails Session details
- * @param sessionSignature Session signature
- * @returns Encoded smart session signature
- */
-function encodeSmartSessionSignature(
-  sessionDetails: SessionDetails,
-  sessionSignature: Hex,
-) {
-  return encodeSmartSessionSignatureInternal(
-    sessionDetails.mode,
-    sessionDetails.enableSessionData.permissionId,
-    sessionSignature,
-    sessionDetails.enableSessionData,
-  )
-}
-
-export { enableSession, encodeSmartSessionSignature }
+export { enableSession }